Adding more tests and cleanup
