+++ src/main/java/formflow/library/file/FileVirusScanner.java
@@ -7,5 +7,5 @@ import org.springframework.web.multipart.MultipartFile;
  */
 public interface FileVirusScanner {
 
-  abstract Boolean doesFileHaveVirus(MultipartFile file) throws Exception;
+  Boolean doesFileHaveVirus(MultipartFile file) throws Exception;
 }
