+++ src/main/resources/messages-form-flow.properties
@@ -43,7 +43,7 @@ upload-documents.error-password-protected=We are unable to process password-prot
 upload-documents.error-could-not-read-file=We could not read the uploaded file. Please try another file.
 upload-documents.error-mime-type=File contents did not match accepted mime types.
 upload-documents.error-virus-found=We are unable to process this file because a virus was detected. Please try another file.
-upload-documents.error-virus-scanner-unavailable=We are unable to process your files at this time. Please try again later.
+upload-documents.error-virus-scanner-unavailable=The upload service could not be reached at this time. Please try again later. 
 #
 demo.banner-text=<b>This site is for example purposes only</b>
 #
